pyproject.toml
[project]
name = "quantum-mini-apps"
version = "0.1.0"
requires-python = ">=3.12"

[tool.poetry]
name = "quantum-mini-apps"
version = "0.1.0"
description = "Framework for Developing Mini Applications for Quantum-HPC Systems"
authors = ["Pradeep Mantha, Nishant Saurabh, Florian Kiwit, Shantenu Jha, Andre Luckow"]
readme = "README.md"
packages = [
     { include = "engine", from = "src" },
     { include = "mini_apps", from = "src" }     
]

[tool.poetry.dependencies]
python = "^3.12"
#qugen = {git = "https://github.com/drelu/qugen"}
pennylane = { version = "*"}
pilot-quantum = {git = "https://github.com/radical-cybertools/pilot-quantum.git"}
"dask[complete]" = "^2025.12.0"
dask-expr = "^2.0.0"
distributed = "^2025.12.0"
ray = {version = "^2.52.0", extras = ["all"]}
qiskit = "^1.2.0"
qiskit-aer = "^0.15.0"
qiskit-addon-cutting = "^0.9.0"
qiskit_ibm_runtime = "^0.24.0"
qiskit_ionq = "^0.6.0"
fire = "*"
qiskit-aer-gpu = {version = "^0.15.0", optional = true}

[tool.poetry.extras]
gpu = ["qiskit-aer-gpu"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
